Remove "[skip ci]" annotations in PR titles
`[skip ci]` is a string that is commonly used on CI/CD systems to
avoid a PR triggering a CI build.

For example, for a readme only PR, there is no need to trigger a new
CI build, so adding `[skip ci]` will avoid this.

Then those annotations are not necessary for changelog purposes